chore(cubesql): ProtocolDetails - support session variables (#8587)
