+++ rust/cubesql/cubesql/src/sql/database_variables/mysql/global_vars.rs
@@ -1,8 +1,7 @@
 use std::collections::HashMap;
 
+use crate::compile::{DatabaseVariable, DatabaseVariables};
 use datafusion::scalar::ScalarValue;
-
-use crate::sql::database_variables::{DatabaseVariable, DatabaseVariables};
 
 pub fn defaults() -> DatabaseVariables {
     let mut variables: DatabaseVariables = HashMap::new();
