+++ rust/cubesql/cubesql/src/sql/database_variables/postgres/session_vars.rs
@@ -1,7 +1,7 @@
 use datafusion::scalar::ScalarValue;
 use std::collections::HashMap;
 
-use crate::sql::database_variables::{DatabaseVariable, DatabaseVariables};
+use crate::compile::{DatabaseVariable, DatabaseVariables};
 
 pub fn defaults() -> DatabaseVariables {
     let mut variables: DatabaseVariables = HashMap::new();
