+++ rust/cubesql/cubesql/src/sql/server_manager.rs
@@ -1,20 +1,15 @@
 use crate::{
-    compile::DatabaseProtocol,
+    compile::{DatabaseProtocol, DatabaseVariables, DatabaseVariablesToUpdate},
     config::ConfigObj,
     sql::{
         compiler_cache::CompilerCache,
-        database_variables::{
-            mysql_default_global_variables, postgres_default_global_variables,
-            DatabaseVariablesToUpdate,
-        },
+        database_variables::{mysql_default_global_variables, postgres_default_global_variables},
         SqlAuthService,
     },
     transport::TransportService,
     CubeError,
 };
 use std::sync::{Arc, RwLock as RwLockSync, RwLockReadGuard, RwLockWriteGuard};
-
-use super::database_variables::DatabaseVariables;
 
 #[derive(Debug)]
 pub struct ServerConfiguration {
